RANGER-1124:Good coding practices
(cherry picked from commit a684b48c5a09b2432017249eefad70bc8e36123a)

+++ security-admin/src/main/java/org/apache/ranger/rest/TagREST.java
@@ -191,7 +191,9 @@ public class TagREST {
 
         try {
             RangerTagDef exist = tagStore.getTagDefByGuid(guid);
-            tagStore.deleteTagDef(exist.getId());
+            if(exist!=null){
+				tagStore.deleteTagDef(exist.getId());
+			}
         } catch(Exception excp) {
             LOG.error("deleteTagDef(" + guid + ") failed", excp);
